requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
attrs==22.2.0
    # via -r requirements.in
beautifulsoup4==4.12.0
    # via -r requirements.in
black==23.1.0
    # via -r requirements.in
blurhash==1.1.4
    # via mastodon-py
build==0.10.0
    # via pip-tools
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
    #   typer
decorator==5.1.1
    # via mastodon-py
idna==3.4
    # via requests
isort==5.12.0
    # via -r requirements.in
markdown-it-py==2.2.0
    # via rich
mastodon-py==1.8.0
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.0.0
    # via black
packaging==23.0
    # via
    #   black
    #   build
pathspec==0.11.1
    # via black
pip-tools==6.12.3
    # via -r requirements.in
platformdirs==3.1.1
    # via black
pygments==2.14.0
    # via rich
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   mastodon-py
python-magic==0.4.27
    # via mastodon-py
requests==2.28.2
    # via
    #   -r requirements.in
    #   mastodon-py
rich==13.3.2
    # via -r requirements.in
six==1.16.0
    # via
    #   mastodon-py
    #   python-dateutil
soupsieve==2.4
    # via beautifulsoup4
typer==0.7.0
    # via -r requirements.in
urllib3==1.26.15
    # via requests
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
